@@ -1,5 +1,5 @@
 import {expectType} from 'tsd';
-import hasYarn = require('.');
+import hasYarn from './index.js';
 
 expectType<boolean>(hasYarn());
 expectType<boolean>(hasYarn('foo'));
